Feat freq shifting

+++ VC_firmware/VC_firmware/src/interface.h
@@ -6,8 +6,8 @@
 #include "config.h"
 
 void send_message(String channel, String mssg, float data);
-int validator(int &user_height, int &user_period, String &input);
-int check_height(int &num);
-int check_period(int &num);
+int validator(float &user_height, float &user_period, String &input);
+int check_height(float &num);
+int check_period(float &num);
 String get_input();
-int listener(int &user_height, int &user_period);
+int listener(float &user_height, float &user_period);
